+++ app/ServiceManagement.swift
@@ -1,0 +1,6 @@
+import CoreFoundation
+import ServiceManagement
+
+func setLoginItem(enabled: Bool) {
+    SMLoginItemSetEnabled("net.bacongravy.giphy-anywhere-helper" as CFString, enabled)
+}
